clippy_config

Module conf

Modules

defaults 🔒

Macros

default_text 🔒

define_Conf 🔒

deserialize 🔒

wrap_option 🔒

Structs

Conf

Clippy lint configuration

ConfError 🔒

ConfVisitor 🔒

FieldError 🔒

Suggestion 🔒

TryConf 🔒

Conf with parse errors

Enums

Field 🔒

Constants

DEFAULT_ALLOWED_IDENTS_BELOW_MIN_CHARS 🔒

DEFAULT_ALLOWED_PREFIXES 🔒

DEFAULT_ALLOWED_TRAITS_WITH_RENAMED_PARAMS 🔒

DEFAULT_DISALLOWED_NAMES 🔒

DEFAULT_DOC_VALID_IDENTS 🔒

DEFAULT_MODULE_ITEM_ORDERING_GROUPS 🔒

DEFAULT_SOURCE_ITEM_ORDERING 🔒

DEFAULT_TRAIT_ASSOC_ITEM_KINDS_ORDER 🔒

SEPARATOR_WIDTH 🔒

Functions

calculate_dimensions 🔒

deserialize 🔒

extend_vec_if_indicator_present 🔒

get_configuration_metadata

lookup_conf_file

Search for the configuration file.

sanitize_explanation

span_from_toml_range 🔒

suggest_candidate 🔒

Given a user-provided value that couldn’t be matched to a known option, finds the most likely candidate among candidates that the user might have meant.

union 🔒