clippy_config

Module types

Macros

unimplemented_serialize 🔒

Structs

DisallowedPath

MacroMatcher

Rename

SourceItemOrdering

Represents which item categories are enabled for ordering.

SourceItemOrderingModuleItemGroupings

Represents the configured ordering of items within a module.

SourceItemOrderingTraitAssocItemKinds

Represents the order in which associated trait items should be ordered.

Enums

DisallowedPathEnum 🔒

MatchLintBehaviour

PubUnderscoreFieldsBehaviour

SourceItemOrderingCategory

Represents the item categories that can be ordered by the source ordering lint.

SourceItemOrderingModuleItemKind

Represents the items that can occur within a module.

SourceItemOrderingTraitAssocItemKind

Represents all kinds of trait associated items.

SourceItemOrderingWithinModuleItemGroupings

Describes which specific groupings should have their items ordered alphabetically.

StringOrVecOfString 🔒

Helper struct for deserializing the SourceItemOrderingWithinModuleItemGroupings.

Functions

create_disallowed_map

Creates a map of disallowed items to the reason they were disallowed.

Type Aliases

DisallowedPathWithoutReplacement