clippy_config::types

Type Alias DisallowedPathWithoutReplacement

pub type DisallowedPathWithoutReplacement = DisallowedPath<false>;

Aliased Type

pub struct DisallowedPathWithoutReplacement {
    path: String,
    reason: Option<String>,
    replacement: Option<String>,
    allow_invalid: bool,
    span: Span,
}

Fields
path: Stringreason: Option<String>replacement: Option<String>allow_invalid: bool

Setting allow_invalid to true suppresses a warning if path does not refer to an existing definition.

This could be useful when conditional compilation is used, or when a clippy.toml file is shared among multiple projects.

span: Span

The span of the DisallowedPath.

Used for diagnostics.

Layout

Note: Most layout information is completely unstable and may even differ between compilations. The only exception is types with certain repr(...) attributes. Please see the Rust Reference's “Type Layout” chapter for details on type layout guarantees.

Size: 88 bytes